backend/pyproject.toml
[tool.poetry]
name = "fa-pos-backend"
version = "0.1.0"
description = "FastAPI backend for the FA POS system (Supabase integration)"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.32.0" }
pydantic-settings = "^2.6.1"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.36" }
asyncpg = "^0.29.0"
aiosqlite = "^0.20.0"
databases = "^0.9.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
httpx = "^0.27.2"
supabase = "^2.5.1"
alembic = "^1.13.3"
python-multipart = "^0.0.9"
pydantic = {extras = ["email"], version = "^2.12.4"}
bcrypt = "^5.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
anyio = "^4.6.2"
black = "^24.10.0"
ruff = "^0.6.8"
mypy = "^1.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
